refactor : 가독성을 위해 short이 아닌 int 타입으로 데이터를 다루도록 변경

+++ src/main/java/com/prgmrs/voucher/repository/VoucherRepository.java
@@ -18,7 +18,7 @@ public interface VoucherRepository {
 
     List<Voucher> getAssignedVoucherList();
 
-    List<Voucher> findByCreationTimeAndDiscountType(LocalDateTime startDate, LocalDateTime endDate, short discountType);
+    List<Voucher> findByCreationTimeAndDiscountType(LocalDateTime startDate, LocalDateTime endDate, int discountType);
 
     int removeVoucher(UUID voucherUUID);
